+++ src/js/hooks/useColorTheme.ts
@@ -11,14 +11,12 @@ import { decimalMultiplier, decimalToHex, sendToElectron } from 'js/utils';
  */
 
 function useColorTheme(): void {
-  const defaultTheme = 'chromatix';
+  const defaultTheme = 'chromatix-magenta';
 
   const dispatch = useDispatch();
 
   const fullPageMode = useSelector(({ appModel }: any) => appModel.fullPageMode);
   const fullPageTheme = useSelector(({ sessionModel }: any) => sessionModel.fullPageTheme);
-
-  const accessibilityContrast = useSelector(({ sessionModel }: any) => sessionModel.accessibilityContrast);
 
   const currentServer = useSelector(({ sessionModel }: any) => sessionModel.currentServer);
   const currentLibrary = useSelector(({ sessionModel }: any) => sessionModel.currentLibrary);
@@ -28,10 +26,11 @@ function useColorTheme(): void {
   const hasQueueVisible = queueIsVisible && hasSelectedLibrary;
 
   const currentTheme = useSelector(({ sessionModel }: any) => sessionModel.currentTheme);
-
   const currentColorBackground = useSelector(({ sessionModel }: any) => sessionModel.currentColorBackground);
   const currentColorPrimary = useSelector(({ sessionModel }: any) => sessionModel.currentColorPrimary);
   const currentColorText = useSelector(({ sessionModel }: any) => sessionModel.currentColorText);
+  const currentUiTinting = useSelector(({ sessionModel }: any) => sessionModel.currentUiTinting);
+  const currentContrast = useSelector(({ sessionModel }: any) => sessionModel.currentContrast);
 
   useEffect(() => {
     const actualTheme =
@@ -44,99 +43,167 @@ function useColorTheme(): void {
     const colorCore =
       currentTheme === 'custom' ? currentColorPrimary : themes[actualTheme as keyof typeof themes].primary;
     const colorText = currentTheme === 'custom' ? currentColorText : themes[actualTheme as keyof typeof themes].text;
-    const colorPrimaryBackground =
+    const colorPrimaryBg =
       currentTheme === 'custom' ? currentColorBackground : themes[actualTheme as keyof typeof themes].background;
 
-    const chromaMultiplier = accessibilityContrast ? 1.4 : 1.1;
-    const opacityMultiplier = accessibilityContrast ? 1.6 : 1;
+    const contrastLevel = currentContrast === 'high' ? 'high' : currentContrast === 'medium' ? 'medium' : 'default';
 
-    const isLightTheme = chroma(colorPrimaryBackground).luminance() > 0.5;
+    const chromaMultiplierLow = contrastLevel === 'high' ? 1.5 : contrastLevel === 'medium' ? 1.3 : 1.1;
+    const chromaMultiplierHigh = contrastLevel === 'high' ? 1.7 : contrastLevel === 'medium' ? 1.4 : 1.1;
+    const opacityMultiplier = contrastLevel === 'high' ? 2 : contrastLevel === 'medium' ? 1.5 : 1;
 
-    let colorBlackout;
-    let colorButtonSecondary;
+    const primaryBackgroundChroma = chroma(colorPrimaryBg);
 
-    let colorSecondaryBackground: string;
+    // [NOTE] Old method
+    // const isLightTheme = primaryBackgroundChroma.luminance() > 0.6;
+    const isLightTheme =
+      chroma.contrast(primaryBackgroundChroma.hex(), '#000000') >=
+      chroma.contrast(primaryBackgroundChroma.hex(), '#ffffff');
+
+    const isLightText =
+      chroma.contrast(chroma(colorText).hex(), '#000000') >= chroma.contrast(chroma(colorText).hex(), '#ffffff');
+
+    const isActualLightTheme =
+      currentTheme === 'custom'
+        ? currentUiTinting === 'darken' || (currentUiTinting === 'auto' && isLightTheme)
+        : isLightTheme;
+
+    let colorBlackout: string;
+
+    let colorSecondaryBg: string;
     let colorSecondaryBorder: string;
-    let colorSecondaryCard: string;
     let colorSecondaryHover: string;
     let colorSecondaryActive: string;
 
-    let colorTertiaryBackground: string;
+    let colorTertiaryBg: string;
     let colorTertiaryBorder: string;
     let colorTertiaryHover: string;
 
-    // Light theme handling
-    if (isLightTheme) {
+    let colorCardOnPrimaryBg: string;
+    let colorCardOnPrimaryBorder: string;
+
+    let colorCardOnSecondaryBg: string;
+    let colorCardOnSecondaryBorder: string;
+
+    let colorButtonMonoBg: string;
+
+    // [NOTE] Optional tinting maybe to be added in future
+
+    // // Tint secondary and tertiary colours using the core colour
+    // const [red, green, blue] = primaryBackgroundChroma.rgb();
+    // const isNeutralBackground = Math.max(red, green, blue) - Math.min(red, green, blue) <= 2;
+    // const shouldApplyCoreTint = isLightTheme && isNeutralBackground;
+
+    // if (shouldApplyCoreTint) {
+    //   const applyCoreTint = (color: string): string => chroma.mix(color, colorCore, 0.02, 'lab').hex();
+
+    //   colorSecondaryBg = applyCoreTint(colorSecondaryBg);
+    //   colorSecondaryBorder = applyCoreTint(colorSecondaryBorder);
+    //   colorCardOnPrimaryBg = applyCoreTint(colorCardOnPrimaryBg);
+    //   colorSecondaryHover = applyCoreTint(colorSecondaryHover);
+    //   colorSecondaryActive = applyCoreTint(colorSecondaryActive);
+
+    //   colorTertiaryBg = applyCoreTint(colorTertiaryBg);
+    //   colorTertiaryBorder = applyCoreTint(colorTertiaryBorder);
+    //   colorTertiaryHover = applyCoreTint(colorTertiaryHover);
+    // }
+
+    // Light theme handling (new)
+    if (isActualLightTheme) {
       colorBlackout =
-        chroma(colorPrimaryBackground)
-          .darken(0.2 * chromaMultiplier)
+        chroma(colorPrimaryBg)
+          .darken(0.2 * chromaMultiplierHigh)
           .hex() + decimalToHex(decimalMultiplier(opacityMultiplier, 0.9));
-      colorButtonSecondary = chroma(colorText)
-        .brighten(0.3 * chromaMultiplier)
+
+      colorSecondaryBg = chroma(colorPrimaryBg)
+        .darken(0.2 * chromaMultiplierLow)
+        .hex();
+      colorSecondaryBorder = chroma(colorPrimaryBg)
+        .darken(0.6 * chromaMultiplierHigh)
+        .hex();
+      colorSecondaryHover = chroma(colorPrimaryBg)
+        .darken(0.4 * chromaMultiplierHigh)
+        .hex();
+      colorSecondaryActive = chroma(colorPrimaryBg)
+        .darken(0.6 * chromaMultiplierHigh)
         .hex();
 
-      colorSecondaryBackground = chroma(colorPrimaryBackground)
-        .darken(0.4 * chromaMultiplier)
+      colorTertiaryBg = chroma(colorPrimaryBg)
+        .darken(0.6 * chromaMultiplierLow)
         .hex();
-      colorSecondaryBorder = chroma(colorPrimaryBackground)
-        .darken(0.8 * chromaMultiplier)
+      colorTertiaryBorder = chroma(colorPrimaryBg)
+        .darken(0.9 * chromaMultiplierHigh)
         .hex();
-      colorSecondaryCard = chroma(colorPrimaryBackground)
-        .darken(0.55 * chromaMultiplier)
-        .hex();
-      colorSecondaryHover = chroma(colorPrimaryBackground)
-        .darken(0.65 * chromaMultiplier)
-        .hex();
-      colorSecondaryActive = chroma(colorPrimaryBackground)
-        .darken(0.75 * chromaMultiplier)
+      colorTertiaryHover = chroma(colorPrimaryBg)
+        .darken(0.9 * chromaMultiplierHigh)
         .hex();
 
-      colorTertiaryBackground = chroma(colorPrimaryBackground)
-        .darken(0.9 * chromaMultiplier)
+      colorCardOnPrimaryBg = chroma(colorPrimaryBg)
+        .darken(0.2 * chromaMultiplierLow)
         .hex();
-      colorTertiaryBorder = chroma(colorPrimaryBackground)
-        .darken(1.2 * chromaMultiplier)
+      colorCardOnPrimaryBorder = chroma(colorPrimaryBg)
+        .darken(0.6 * chromaMultiplierHigh)
         .hex();
-      colorTertiaryHover = chroma(colorPrimaryBackground)
-        .darken(1.25 * chromaMultiplier)
+
+      colorCardOnSecondaryBg = chroma(colorPrimaryBg)
+        .darken(0.6 * chromaMultiplierLow)
+        .hex();
+      colorCardOnSecondaryBorder = chroma(colorPrimaryBg)
+        .darken(0.6 * chromaMultiplierHigh) // [NOTE] invisible border
+        .hex();
+
+      colorButtonMonoBg = chroma(colorText)
+        .brighten(0.3 * chromaMultiplierHigh)
         .hex();
     }
 
     // Dark theme handling
     else {
       colorBlackout =
-        chroma(colorPrimaryBackground)
-          .darken(0.5 * chromaMultiplier)
+        chroma(colorPrimaryBg)
+          .darken(0.5 * chromaMultiplierHigh)
           .desaturate(0.35)
           .hex() + decimalToHex(decimalMultiplier(opacityMultiplier, 0.8));
-      colorButtonSecondary = chroma(colorText)
-        .darken(0.3 * chromaMultiplier)
+
+      colorSecondaryBg = chroma(colorPrimaryBg)
+        .brighten(0.4 * chromaMultiplierHigh)
+        .hex();
+      colorSecondaryBorder = chroma(colorPrimaryBg)
+        .brighten(0.8 * chromaMultiplierHigh)
+        .hex();
+      colorSecondaryHover = chroma(colorPrimaryBg)
+        .brighten(0.6 * chromaMultiplierHigh)
+        .hex();
+      colorSecondaryActive = chroma(colorPrimaryBg)
+        .brighten(0.75 * chromaMultiplierHigh)
         .hex();
 
-      colorSecondaryBackground = chroma(colorPrimaryBackground)
-        .brighten(0.4 * chromaMultiplier)
+      colorTertiaryBg = chroma(colorPrimaryBg)
+        .brighten(0.9 * chromaMultiplierHigh)
         .hex();
-      colorSecondaryBorder = chroma(colorPrimaryBackground)
-        .brighten(0.8 * chromaMultiplier)
+      colorTertiaryBorder = chroma(colorPrimaryBg)
+        .brighten(1.2 * chromaMultiplierHigh)
         .hex();
-      colorSecondaryCard = chroma(colorPrimaryBackground)
-        .brighten(0.55 * chromaMultiplier)
-        .hex();
-      colorSecondaryHover = chroma(colorPrimaryBackground)
-        .brighten(0.65 * chromaMultiplier)
-        .hex();
-      colorSecondaryActive = chroma(colorPrimaryBackground)
-        .brighten(0.75 * chromaMultiplier)
+      colorTertiaryHover = chroma(colorPrimaryBg)
+        .brighten(1.25 * chromaMultiplierHigh)
         .hex();
 
-      colorTertiaryBackground = chroma(colorPrimaryBackground)
-        .brighten(0.9 * chromaMultiplier)
+      colorCardOnPrimaryBg = chroma(colorPrimaryBg)
+        .brighten(0.55 * chromaMultiplierHigh)
         .hex();
-      colorTertiaryBorder = chroma(colorPrimaryBackground)
-        .brighten(1.2 * chromaMultiplier)
+      colorCardOnPrimaryBorder = chroma(colorPrimaryBg)
+        .brighten(0.85 * chromaMultiplierHigh)
         .hex();
-      colorTertiaryHover = chroma(colorPrimaryBackground)
-        .brighten(1.25 * chromaMultiplier)
+
+      colorCardOnSecondaryBg = chroma(colorPrimaryBg)
+        .brighten(0.9 * chromaMultiplierHigh)
+        .hex();
+      colorCardOnSecondaryBorder = chroma(colorPrimaryBg)
+        .brighten(1.2 * chromaMultiplierHigh)
+        .hex();
+
+      colorButtonMonoBg = chroma(colorText)
+        .darken(0.3 * chromaMultiplierHigh)
         .hex();
     }
 
@@ -154,11 +221,6 @@ function useColorTheme(): void {
     const colorOpacity07 = colorText + decimalToHex(decimalMultiplier(opacityMultiplier, 0.7));
     const colorOpacity08 = colorText + decimalToHex(decimalMultiplier(opacityMultiplier, 0.8));
 
-    const shadowLg = isLightTheme ? '0 4px 20px rgba(0, 0, 0, 0.1)' : '0 2px 20px rgba(0, 0, 0, 0.4)';
-    const shadowMd = isLightTheme ? '0 4px 6px rgba(0, 0, 0, 0.05)' : '0 2px 10px rgba(0, 0, 0, 0.4)';
-    const shadowSm = isLightTheme ? '0 4px 6px rgba(0, 0, 0, 0.05)' : '0 2px 8px rgba(0, 0, 0, 0.35)';
-    const shadowXs = isLightTheme ? '0 2px 4px rgba(0, 0, 0, 0.05)' : '0 2px 6px rgba(0, 0, 0, 0.2)';
-
     const opacity02 = decimalMultiplier(opacityMultiplier, 0.2);
     const opacity025 = decimalMultiplier(opacityMultiplier, 0.25);
     const opacity03 = decimalMultiplier(opacityMultiplier, 0.3);
@@ -168,23 +230,39 @@ function useColorTheme(): void {
     const opacity07 = decimalMultiplier(opacityMultiplier, 0.7);
     const opacity08 = decimalMultiplier(opacityMultiplier, 0.8);
 
+    const shadowLg = isActualLightTheme ? '0 4px 20px rgba(0, 0, 0, 0.1)' : '0 2px 20px rgba(0, 0, 0, 0.35)';
+    const shadowMd = isActualLightTheme ? '0 4px 8px rgba(0, 0, 0, 0.1)' : '0 2px 10px rgba(0, 0, 0, 0.35)';
+    const shadowSm = isActualLightTheme ? '0 2px 4px rgba(0, 0, 0, 0.05)' : '0 2px 6px rgba(0, 0, 0, 0.2)';
+
     const colors: Record<string, string | number> = {
       '--color-core': colorCore,
       '--color-text': colorText,
-      '--color-primary-background': colorPrimaryBackground,
       '--color-blackout': colorBlackout,
 
-      '--color-button-secondary': colorButtonSecondary,
+      '--color-primary-bg': colorPrimaryBg,
 
-      '--color-secondary-background': colorSecondaryBackground,
+      '--color-secondary-bg': colorSecondaryBg,
       '--color-secondary-border': colorSecondaryBorder,
-      '--color-secondary-card': colorSecondaryCard,
       '--color-secondary-hover': colorSecondaryHover,
       '--color-secondary-active': colorSecondaryActive,
 
-      '--color-tertiary-background': colorTertiaryBackground,
+      '--color-tertiary-bg': colorTertiaryBg,
       '--color-tertiary-border': colorTertiaryBorder,
       '--color-tertiary-hover': colorTertiaryHover,
+
+      '--color-card-on-primary-bg': colorCardOnPrimaryBg,
+      '--color-card-on-primary-border': colorCardOnPrimaryBorder,
+
+      '--color-card-on-secondary-bg': colorCardOnSecondaryBg,
+      '--color-card-on-secondary-border': colorCardOnSecondaryBorder,
+
+      '--color-button-on-primary-bg': colorCardOnPrimaryBg,
+      '--color-button-on-primary-border': colorCardOnPrimaryBorder,
+
+      '--color-button-on-secondary-bg': colorTertiaryBg,
+      '--color-button-on-secondary-border': colorTertiaryBorder,
+
+      '--color-button-mono-bg': colorButtonMonoBg,
 
       '--color-opacity-0025': colorOpacity0025,
       '--color-opacity-005': colorOpacity005,
@@ -200,11 +278,6 @@ function useColorTheme(): void {
       '--color-opacity-07': colorOpacity07,
       '--color-opacity-08': colorOpacity08,
 
-      '--shadow-lg': shadowLg,
-      '--shadow-md': shadowMd,
-      '--shadow-sm': shadowSm,
-      '--shadow-xs': shadowXs,
-
       '--opacity-02': opacity02,
       '--opacity-025': opacity025,
       '--opacity-03': opacity03,
@@ -213,6 +286,10 @@ function useColorTheme(): void {
       '--opacity-06': opacity06,
       '--opacity-07': opacity07,
       '--opacity-08': opacity08,
+
+      '--shadow-lg': shadowLg,
+      '--shadow-md': shadowMd,
+      '--shadow-sm': shadowSm,
     };
 
     // Get or create the style element for dynamic theme variables
@@ -231,19 +308,21 @@ function useColorTheme(): void {
 
     // Send updated colors to Electron main process
     sendToElectron(['win', 'lin'], 'color-theme', {
-      background: hasQueueVisible && !fullPageMode ? colorSecondaryBackground : colorPrimaryBackground,
+      background: hasQueueVisible && !fullPageMode ? colorSecondaryBg : colorPrimaryBg,
       text: colorText,
       primary: colorCore,
     });
 
     // Save whether the theme is light or dark in the session state
     dispatch.sessionModel.setSessionState({
-      isLightTheme,
+      isLightTheme: isActualLightTheme,
+      isLightText: isLightText,
     });
 
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [
-    accessibilityContrast,
+    currentContrast,
+    currentUiTinting,
     hasQueueVisible,
     currentTheme,
     currentColorBackground,