+++ src/js/pages/DevSettings.jsx
@@ -1,0 +1,41 @@
+// ======================================================================
+// IMPORTS
+// ======================================================================
+
+import {
+  SettingsAbout,
+  SettingsAppearance,
+  SettingsBrowse,
+  SettingsChangelog,
+  SettingsControls,
+  SettingsDownloads,
+  SettingsGeneral,
+  SettingsSidebar,
+  TitleHeading,
+} from 'js/components';
+
+// ======================================================================
+// COMPONENT
+// ======================================================================
+
+const Component = () => {
+  return (
+    <>
+      <TitleHeading title="All Settings" />
+      <SettingsGeneral debug={true} />
+      <SettingsBrowse debug={true} />
+      <SettingsAppearance debug={true} />
+      <SettingsSidebar />
+      <SettingsControls />
+      <SettingsDownloads />
+      <SettingsAbout />
+      <SettingsChangelog />
+    </>
+  );
+};
+
+// ======================================================================
+// EXPORT
+// ======================================================================
+
+export default Component;